+++ packages/yourloops/test/integration/assert/drawer.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Diabeloop
+ * Copyright (c) 2022-2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -32,7 +32,6 @@ export const checkDrawer = () => {
   expect(drawer.getByLabelText('Filter on all patients')).toBeVisible()
   expect(drawer.getByLabelText('Filter on flagged patients')).toBeVisible()
   expect(drawer.getByLabelText('Filter on pending patients')).toBeVisible()
-  expect(drawer.getByLabelText('Filter on patients in private practice')).toBeVisible()
 }
 
 export const checkDrawerNotVisible = () => {